+++ edgy/cli/operations/edit.py
@@ -1,12 +1,13 @@
-import click
+import sayer
 
 from edgy.cli.base import edit as _edit
 from edgy.cli.decorators import add_migration_directory_option
 
+from ..common_params import RevisionHeadArgument
+
 
 @add_migration_directory_option
-@click.command()
-@click.argument("revision", default="head")
-def edit(revision: str) -> None:
+@sayer.command
+def edit(revision: RevisionHeadArgument) -> None:
     """Edit a revision file"""
     _edit(revision)