+++ edgy/cli/operations/shell/enums.py
@@ -4,4 +4,4 @@ from enum import Enum
 class ShellOption(str, Enum):
     IPYTHON = "ipython"
     PTPYTHON = "ptpython"
-    PYTHON = "python"
+    # PYTHON = "python"