+++ edgy/contrib/admin/views.py
@@ -4,6 +4,7 @@ from typing import TYPE_CHECKING, Any, cast
 
 import anyio
 import orjson
+from lilya import status
 from lilya.controllers import Controller
 from lilya.exceptions import NotFound
 from lilya.requests import Request
@@ -12,7 +13,7 @@ from lilya.templating.controllers import TemplateController
 from pydantic import ValidationError
 
 import edgy
-from edgy.contrib.admin.mixins import AdminMixin
+from edgy.contrib.admin.mixins import AdminMixin, get_templates
 from edgy.contrib.admin.utils.messages import add_message
 from edgy.contrib.pagination import Paginator
 from edgy.core.db.fields.file_field import ConcreteFileField
@@ -870,3 +871,24 @@ class ModelObjectCreateView(BaseObjectView, AdminMixin, TemplateController):
         return RedirectResponse(
             f"{self.get_admin_prefix_url(request)}/models/{model_name}/{obj_id}"
         )
+
+
+async def admin_not_found(request: Request, exc: Exception) -> Any:
+    """
+    An asynchronous exception handler for 404 Not Found errors.
+
+    This function renders a 404.html.jinja template and returns it with a 404 status code.
+
+    Args:
+        request (Request): The incoming Lilya request object.
+        exc (Exception): The exception that triggered this handler (e.g., NotFound).
+
+    Returns:
+        Any: A Jinja2TemplateResponse object for the 404 page.
+    """
+    return get_templates().get_template_response(
+        request,
+        "admin/404.html.jinja",
+        context={"title": "Not Found"},
+        status_code=status.HTTP_404_NOT_FOUND,
+    )