+++ docs/release-notes.md
@@ -13,6 +13,12 @@ hide:
 - For `ManyToManyField` make `NEW_M2M_NAMING` the default. No need to set it explicit anymore.
 - Use `sayer` for cli.
 - Add .jinja suffix to admin templates (makes it easier to recognize jinja templates for IDEs and humans).
+- Move `404.html.jinja` to admin, as it is solely used by `admin_serve` now.
+
+### Fixed
+
+- The `create_admin_app` created lilya instance doesn't intercept 404 errors anymore.
+- `404.html.jinja` uses now also the `base_definitions.html.jinja` so we can exchange libraries at one place.
 
 ### Warning