+++ docs/release-notes.md
@@ -6,6 +6,13 @@ hide:
 
 # Release Notes
 
+## 0.34.1
+
+### Added
+
+- [add_many()](./queries/many-to-many.md#add_many) to ManyToManyField for adding multiple relations at once.
+- [remove_many()](./queries/many-to-many.md#remove_many) to ManyToManyField for removing multiple relations at once.
+
 ## 0.34.0
 
 ### Changed